requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.12.0
    # via ortools
aiohttp==3.7.4.post0
    # via discord.py
asgiref==3.3.4
    # via django
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
cryptography==3.4.7
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
discord.py==1.7.1
    # via -r requirements.in
django-allauth==0.44.0
    # via -r requirements.in
django-debug-toolbar==3.2
    # via -r requirements.in
django-tailwind==1.2.0
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-tailwind
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==2.10
    # via
    #   requests
    #   yarl
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.1.0
    # via requests-oauthlib
ortools==8.2.8710
    # via -r requirements.in
protobuf==3.15.8
    # via ortools
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.0.1
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via django
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   django-allauth
    #   requests-oauthlib
six==1.15.0
    # via
    #   absl-py
    #   protobuf
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==3.7.4.3
    # via aiohttp
urllib3==1.26.4
    # via requests
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
